Move common sorting code to base Transformer class
